+++ combust/CHANGELOG.md
@@ -4,7 +4,7 @@ The following tools were used to generate this project:
 - Angular CLI (ng)
 
 The following steps were used to generate this project:
-- Create Angular project with ng: `ng new combust --defaults --skip-install --skip-git --no-standalone `.
+- Create Angular project with ng: `ng new combust --defaults --skip-install --skip-git `.
 - Create project file (`combust.esproj`).
 - Create `launch.json` to enable debugging.
 - Create `nuget.config` to specify location of the JavaScript Project System SDK (which is used in the first line in `combust.esproj`).